FileGenerator 클래스 역할별 분리 #378
